requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
circuitbreaker==1.4.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
cryptography==38.0.4
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
icmplib==3.0.3
idna==3.4
mysql-connector-python==8.0.31
oci==2.90.1
protobuf==3.20.1
pycparser==2.21
pydantic==1.10.4
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.2
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
vultr==1.0.1
watchfiles==0.18.1
websockets==10.4
